Extract SemanticDrawable abstract class from Strain
The drawable annotations will also be semantic, so we can extract and reuse Strain's implementation.
