+++ dnainator-javafx/src/main/java/nl/tudelft/dnainator/javafx/drawables/strains/Strain.java
@@ -8,6 +8,7 @@ import nl.tudelft.dnainator.core.SequenceNode;
 import nl.tudelft.dnainator.core.impl.Cluster;
 import nl.tudelft.dnainator.graph.Graph;
 import nl.tudelft.dnainator.javafx.ColorServer;
+import nl.tudelft.dnainator.javafx.drawables.SemanticDrawable;
 
 import java.util.HashMap;
 import java.util.List;
@@ -18,20 +19,18 @@ import java.util.stream.Collectors;
  * The {@link Strain} class represents the graph that contains the DNA strain.
  * It holds both content and children, and toggles what to load and display based on the zoom level.
  */
-public class Strain extends Group {
-	private static final int Y_OFFSET = 10;
-	private static final int RANK_WIDTH = 10;
-	private static final int NO_CLUSTERS = 33000;
+public class Strain extends SemanticDrawable {
 	/* JavaFX scene graph cannot handle rectangles larger than 10k pixels, so we split a 30k
 	* rectangle into 4 slices. */
 	private static final int SLICES = 4;
+	private static final int Y_OFFSET = 10;
+	private static final int RANK_WIDTH = 10;
+	private static final int NO_CLUSTERS = 33000;
 	private static final int CLUSTER_DIVIDER = 100;
 
 	private ColorServer colorServer;
 	private Graph graph;
 	private Map<String, ClusterDrawable> clusters;
-	private Group content;
-	private Group childContent;
 
 	/**
 	 * Construct a new top level {@link Strain} using the specified graph.
@@ -51,30 +50,18 @@ public class Strain extends Group {
 	 * @param childContent The specified child content.
 	 */
 	public Strain(ColorServer colorServer, Graph graph, Group content, Group childContent) {
+		super(content, childContent);
 		this.colorServer = colorServer;
 		this.graph = graph;
 		this.clusters = new HashMap<>();
-		this.content = content;
-		this.childContent = childContent;
-
-		getChildren().add(content);
-		getChildren().add(childContent);
 		loadContent();
-	}
-
-	/**
-	 * Check whether this object intersects with the given viewport bounds.
-	 * @param bounds	The given viewport bounds.
-	 * @return	True if (partially) in viewport, false otherwise.
-	 */
-	public boolean isInViewport(Bounds bounds) {
-		return bounds.contains(content.localToParent(0, 0));
 	}
 
 	/*
 	 * Load the drawable content of the graph itself.
 	 */
-	private void loadContent() {
+	@Override
+	protected void loadContent() {
 		for (int i = 0; i < SLICES; i++) {
 			int width = NO_CLUSTERS * RANK_WIDTH / SLICES;
 			Rectangle rectangle = new Rectangle(width, RANK_WIDTH, Color.BLACK);
@@ -83,10 +70,8 @@ public class Strain extends Group {
 		}
 	}
 
-	/*
-	 * Load the drawable content of the graph's children, given the bounds of the viewing port.
-	 */
-	private void loadChildren(Bounds bounds) {
+	@Override
+	protected void loadChildren(Bounds bounds) {
 		int minRank = (int) (Math.max(bounds.getMinX() / RANK_WIDTH, 0));
 		int maxRank = (int) (RANK_WIDTH + bounds.getMaxX() / RANK_WIDTH);
 
@@ -129,38 +114,6 @@ public class Strain extends Group {
 			cluster.setTranslateX(key * RANK_WIDTH);
 			cluster.setTranslateY(i * RANK_WIDTH - value.size() * RANK_WIDTH / 2);
 			childContent.getChildren().add(cluster);
-		}
-	}
-
-	/**
-	 * Toggle between displaying own content or children.
-	 * @param bounds     The bounds of the viewport.
-	 * @param visible    True if the content needs to be visible, false otherwise.
-	 */
-	public void toggle(Bounds bounds, boolean visible) {
-		if (visible && !content.isVisible()) {
-			childContent.getChildren().clear();
-			content.setVisible(true);
-			loadContent();
-		}
-		if (!visible) {
-			content.setVisible(false);
-			loadChildren(bounds);
-		}
-	}
-
-	/**
-	 * Update method that should be called after scaling.
-	 * This method checks how zoomed in we are by transforming bounds to root coordinates,
-	 * and then dynamically adds and deletes items in the JavaFX scene graph.
-	 * @param bounds	The bounds of the viewport to update.
-	 * @param curZoom   The current zoom level.
-	 */
-	public void update(Bounds bounds, double curZoom) {
-		if (curZoom < Thresholds.GRAPH.get()) {
-			toggle(bounds, true);
-		} else {
-			toggle(bounds, false);
 		}
 	}
 }
